feat(tree-probability): update layout and animation, create verifys on choice event and create alerts

+++ src/components/global/Alerts.tsx
@@ -15,13 +15,13 @@ export function Alerts({
 }: Readonly<AlertsProps>) {
 
   useEffect(() => {
-    if(alerts.length > 0 && alerts.every(alert => alert.status === "remove")) {
+    if(alerts && alerts.length > 0 && alerts.every(alert => alert.status === "remove")) {
       deleteAlerts();
     }
   }, [alerts]);
 
   return (
-    alerts.length > 0 && (
+    alerts && alerts.length > 0 && (
       <div className='flex flex-col fixed top-[20vh] right-[16px] items-end z-2 transition-[transform] duration-500 ease-in-out'>
         {alerts.map((alert, index) => {
           return <Alert
